Получение,обработка и отправка данных на сервер
