+++ js/picture-form.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line no-use-before-define
-export {scaleControlValue,imageUploadPreview};
+export {scaleControlValue,imageUploadPreview,slider};
 
 const scaleControlSmaller = document.querySelector('.scale__control--smaller');
 const scaleControlBigger = document.querySelector('.scale__control--bigger');
